pyproject.toml
[project]
name = "ml_bot"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{name = "skip"}]

dependencies = [
    "pandas>=2.1.1",
    "tensorboard>=2.14.1",
    "plotly>=5.17.0",
    "tqdm>=4.66.1",
    "scikit-learn>=1.3.2",
    "matplotlib>=3.8.2",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.1",
    "japanize-matplotlib>=1.1.3",
    "ccxt>=4.2.95",
    "ta-lib>=0.4.28",
    "tensorflow>=2.16.1",
    "keras>=3.3.2",
    "optuna>=3.6.1",
    "optuna-dashboard>=0.15.1",
    "backtrader>=1.9.78.123",
    "rich>=13.8.1",
    "lightgbm>=4.5.0",
    "optuna-integration>=4.0.0",
    "statsmodels>=0.14.4",
    "seaborn>=0.13.2",
    "numerapi>=2.20.0",
    "catboost>=1.2.7",
    "polars>=1.29.0",
    "vectorbt[full]>=0.27.3",
    "pyarrow>=20.0.0",
    "tab-transformer-pytorch>=0.4.2",
]
